requirements.txt
Django>=4.2
djangorestframework~=3.15.2
djangorestframework-simplejwt~=5.3.1
uvicorn[standard]
psycopg2-binary
django-environ~=0.11.2
drf-spectacular~=0.27.2